# swagger/store.swagger.yaml
openapi: 3.0.0
info:
  title: Store API
  version: 1.0.0
  description: "express를 사용한 Store API, 리뷰 미리보기 및 미션 관리 포함"
paths:
  /stores/{storeId}/reviews:
    get:
      summary: 가게의 리뷰 미리보기 조회
      tags:
        - Store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        "200":
          description: 성공적인 응답
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        user_name:
                          type: string
                        rate:
                          type: integer
                        review_content:
                          type: string
                        created_at:
                          type: string
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 가게를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: 내부 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /stores/{storeId}/missions:
    get:
      summary: 특정 가게의 미션 조회
      tags:
        - Store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 성공적인 응답
          content:
            application/json:
              schema:
                type: object
                properties:
                  missions:
                    type: array
                    items:
                      type: object
                      properties:
                        mission_id:
                          type: integer
                        description:
                          type: string
        "404":
          description: 가게를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: 내부 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /stores/{storeId}/users/{userId}/missions:
    get:
      summary: 특정 가게에서 특정 사용자의 미션 조회
      tags:
        - Store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 성공적인 응답
          content:
            application/json:
              schema:
                type: object
                properties:
                  missions:
                    type: array
                    items:
                      type: object
                      properties:
                        mission_id:
                          type: integer
                        description:
                          type: string
                        status:
                          type: string
        "404":
          description: 사용자 또는 가게를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: 내부 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /stores/{storeId}/users/{userId}/missions/{missionId}/complete:
    post:
      summary: 특정 가게에서 특정 사용자의 미션 완료 처리
      tags:
        - Store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: missionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 성공적인 응답
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: 사용자, 가게 또는 미션을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: 내부 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Error:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string